+++ src/main/java/io/elice/shoppingmall/product/Entity/Item/ItemImages.java
@@ -2,7 +2,6 @@ package io.elice.shoppingmall.product.Entity.Item;
 
 import io.elice.shoppingmall.product.DTO.ItemDTO;
 import io.elice.shoppingmall.product.DTO.ItemImagesDTO;
-import io.elice.shoppingmall.product.Entity.Option.Brand;
 import io.elice.shoppingmall.product.Entity.Option.Gender;
 import jakarta.persistence.*;
 import lombok.*;
@@ -17,7 +16,7 @@ public class ItemImages {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column
     private String mainImage;
