+++ src/main/java/io/elice/shoppingmall/product/Repository/ItemImagesRepository.java
@@ -3,5 +3,5 @@ package io.elice.shoppingmall.product.Repository;
 import io.elice.shoppingmall.product.Entity.Item.ItemImages;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemImagesRepository extends JpaRepository<ItemImages, Integer> {
+public interface ItemImagesRepository extends JpaRepository<ItemImages, Long> {
 }
